requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
phonenumbers==8.12.49
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
